delete extra require string and reek disable
