+++ 3741/3/Gemfile
@@ -1,10 +1,9 @@
-# frozen_string_literal: true
-
 source 'https://rubygems.org'
 
 gem 'activerecord'
 gem 'bcrypt'
-gem 'dotenv'
+gem 'dotenv', require: 'dotenv/load'
+gem 'pry'
 gem 'rake'
 gem 'sinatra-activerecord'
 gem 'sinatra-contrib'
