+++ 3741/3/config.ru
@@ -3,4 +3,7 @@
 $LOAD_PATH.unshift '.'
 require 'config/environment'
 
+use SessionsController
+use PlacesController
+use ReviewsController
 run MainController
