+++ 3741/3/controllers/places_controller.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require_relative 'base_controller'
+require_relative 'main_controller'
 
-class PlacesController < BaseController
+class PlacesController < MainController
   namespace '/places' do
     get '/:id' do
       @place = Place.find_by(id: params[:id])
