+++ 3741/3/controllers/sessions_controller.rb
@@ -1,8 +1,8 @@
 # frozen_string_literal: true
 
-require_relative 'base_controller'
-# :reek:InstanceVariableAssumption
-class SessionsController < BaseController
+require_relative 'main_controller'
+
+class SessionsController < MainController
   helpers do
     def login_in?
       return unless user_logged?
@@ -17,13 +17,15 @@ class SessionsController < BaseController
       show_message 'Logout already'
       redirect '/'
     end
-
-    def register_data_valid?
-      @user.valid? && Truemail.valid?(params['email'])
-    end
   end
 
   namespace '/session' do
+    get '/logout' do
+      login_out?
+      session.clear
+      redirect '/'
+    end
+
     get '/signup' do
       login_in?
       erb :signup
@@ -34,43 +36,12 @@ class SessionsController < BaseController
       erb :login
     end
 
-    get '/logout' do
-      login_out?
-      session.clear
-      redirect '/'
-    end
-
     post '/signup' do
-      login_in?
-      @user = User.new(params.slice('username', 'email', 'password'))
       sign_up_redirect
     end
 
     post '/login' do
-      login_in?
-      @user = User.find_by(email: params[:email])
-      login_redirect
-    end
-  end
-
-  def sign_up_redirect
-    if register_data_valid?
-      @user.save
-      show_message 'user create, login in please'
-      redirect '/'
-    else
-      show_message 'email or login already exist'
-      redirect '/session/signup'
-    end
-  end
-
-  def login_redirect
-    if @user && BCrypt::Password.new(@user.password_hash) == params[:password]
-      session[:user_id] = @user.id
-      redirect '/'
-    else
-      show_message 'wrong email'
-      redirect '/session/login'
+      sign_in_redirect
     end
   end
 end
